# [Improvement][Vectorized] Lower/upper case tranfer function vertorized