--- a/be/src/exprs/string_functions.cpp
+++ b/be/src/exprs/string_functions.cpp
@@ -16,6 +16,8 @@
 // under the License.
 
 #include "exprs/string_functions.h"
+#include "util/vectorized-tool/lower.h"
+#include "util/vectorized-tool/upper.h"
 
 #include <re2/re2.h>
 
@@ -346,15 +348,11 @@ StringVal StringFunctions::lower(FunctionContext* context, const StringVal& str)
     if (str.is_null) {
         return StringVal::null();
     }
-    // TODO pengyubing
-    // StringVal result = StringVal::create_temp_string_val(context, str.len);
     StringVal result(context, str.len);
     if (UNLIKELY(result.is_null)) {
         return result;
     }
-    for (int i = 0; i < str.len; ++i) {
-        result.ptr[i] = ::tolower(str.ptr[i]);
-    }
+    Lower::to_lower(str.ptr, str.len, result.ptr);
     return result;
 }
 
@@ -362,15 +360,11 @@ StringVal StringFunctions::upper(FunctionContext* context, const StringVal& str)
     if (str.is_null) {
         return StringVal::null();
     }
-    // TODO pengyubing
-    // StringVal result = StringVal::create_temp_string_val(context, str.len);
     StringVal result(context, str.len);
     if (UNLIKELY(result.is_null)) {
         return result;
     }
-    for (int i = 0; i < str.len; ++i) {
-        result.ptr[i] = ::toupper(str.ptr[i]);
-    }
+    Upper::to_upper(str.ptr, str.len, result.ptr);
     return result;
 }