--- a/be/src/util/vectorized-tool/lower.h
+++ b/be/src/util/vectorized-tool/lower.h
@@ -1,0 +1,37 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+#pragma once
+#ifndef BE_LOWER_H
+#define BE_LOWER_H
+
+#include "lower_upper_impl.h"
+#include <stdint.h>
+
+namespace doris {
+class Lower {
+public:
+    static void to_lower(uint8_t * src, int64_t len, uint8_t * dst) {
+        if (len <= 0) {
+            return;
+        }
+        LowerUpperImpl<'A', 'Z'> lowerUpper;
+        lowerUpper.transfer(src, src + len, dst);
+    }
+};
+}
+
+#endif //BE_LOWER_H