@xinghuayu007 2021-07-19
> This pr refers to Clickhouse's Function implementation.
> So I think it's better to indicate the source.
> Then explain whether you have done any your own optimization or some customized for Doris's case.
> You can see `bitmap_value.h` 's comment.

done